Button locker, spend energy on start game, restart game
Created button locker, which lock button and show energy to use button. Added the logic to spend energy to start or restart game level.

+++ Arcanoid Maket/Assets/Scripts/GameComponents/Energy/Commands/SpendEnergyCommand.cs
@@ -6,7 +6,10 @@ namespace GameComponents.Energy.Commands
     {
         public override void Execute()
         {
-            _service.RemoveEnergy(_energyValue);
+            if (_service.IsEnoughEnergy(_energyValue))
+            {
+                _service.RemoveEnergy(_energyValue);
+            }
         }
     }
 }
